jesus_commandments_website/requirements.txt
# Application core
django>=4.2.4
pandas>=2.0.3
Pillow>=10.0.0
django-vinaigrette>=2.0.1
django-rosetta>=0.9.9
-e ../bible_lib
django-url-or-relative-url-field>=0.2.0
sorl-thumbnail>=12.9.0
mysqlclient>=2.1.0
django-cookie-law>=2.2.0
django-reversion>=5.0.4
whitenoise>=6.5.0

# Tools
-e ../import_tool
-e ../translate_tool
polib>=1.2.0

# Monitoring
sentry-sdk>=1.29.2

# Debug and Development
django-debug-toolbar

# Used to communicatie with youtube
requests
google-api-python-client
oauth2client
